# src/main/java/com/codetets/atm/model/DispenseResponse.java
package com.codetets.atm.model;

public class DispenseResponse {
    private boolean success;
    private String message;

    public DispenseResponse(boolean success, String message) {
        this.success = success;
        this.message = message;
    }

    public boolean isSuccess() {
        return success;
    }

    public void setSuccess(boolean success) {
        this.success = success;
    }

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }
}